Move playlist/liked endpoint  to its own controller
