+++ frontend/src/app/models/album/album.ts
@@ -3,6 +3,7 @@ import { User } from '../user/user';
 export interface Album {
   id: number;
   name: string;
+  description: string;
   iconURL: string;
   artist: User;
   releaseYear: number;
