+++ frontend/src/app/services/playlist-service.service.ts
@@ -5,16 +5,12 @@ import { HttpInternalService } from './http-internal.service';
 @Injectable({
   providedIn: 'root'
 })
-export class PlaylistServiceService {
+export class PlaylistService {
   private _routePrefix = '/api/playlists';
 
   constructor(private _httpService: HttpInternalService) { }
 
   public getAllPlaylists() {
     return this._httpService.getRequest<Playlist[]>(`${this._routePrefix}`);
-  }
-
-  public getLikedPlaylistsByTheUser(userId: number) {
-    return this._httpService.getRequest<Playlist[]>(`${this._routePrefix}/liked/${userId}`);
   }
 }
