+++ frontend/src/app/services/reaction.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
 import { AlbumForReadDTO } from '../models/album/albumForReadDTO';
+import { Playlist } from '../models/playlist';
 import { ArtistReadDTO } from '../models/user/ArtistReadDTO';
 import { HttpInternalService } from './http-internal.service';
 
@@ -61,5 +62,9 @@ export class ReactionService {
       '/api/AlbumReaction/removeLike',
       { albumId, userId }
     );
+  }
+
+  getLikedPlaylistsByTheUser(userId: number) {
+    return this.httpService.getRequest<Playlist[]>(`api/PlaylistReaction/liked/${userId}`);
   }
 }
